+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * change: Sequencex's joinTo() method now throws IllegalStateException
 * new: Collectionx added list, windowed method
 * fix: Fix Collectionx's chunked method bug
+* new: Sequencex added chunked, windowed method
 
 
 ## v1.0.0-rc01